+++ fjs/emergent_testing/module.f.ts
@@ -407,7 +407,7 @@ export const main: NodeProgram =
  * based on the detected `engine`.
  */
 export const register: NodeProgram = o => {
-    const star = o.engine === 'bun' || o.engine === 'playwright' ? ' ...' : ''
+    const star = o.inlineTestContext ? ' ...' : ''
     const ctx = o.engine === 'bun' ? o.bunTestContext :
         o.engine === 'playwright' ? o.playwrightTestContext :
         o.testContext